+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - CHANGELOG!
 - whois command
+- conf_dump command
 - depth to get command
 - lookup net by asn, ixnets to client